requirements.txt
streamlit==1.56.0
pandas==3.0.2
numpy==2.4.4
ultralytics==8.4.34
paho-mqtt==2.1.0
requests==2.33.1
opencv-python-headless==4.10.0.84
streamlit-webrtc==0.47.9
av==12.3.0
twilio==9.4.2
